requirements.txt
Flask
Flask[async]
flask_cors
google-generativeai
PyPDF2
python-dotenv
apscheduler
pytz